[iOS] style: 페어 프로그래밍 코딩 컨벤션 적용
Issue #21
