+++ iOS/AirbnbProject/AirbnbProject/View/MainView/ThumbnailImageCollectionViewCell.xib
